# docs/API_SPEC.yaml
openapi: 3.0.3
info:
  title: Prospect API
  version: 0.1.0
  description: Evidence-first financial document intelligence API.
servers:
  - url: /api/v1
security:
  - sessionToken: []

# Every non-2xx response body is an Error. A document owned by another user returns the
# same 404 as a nonexistent one.
paths:
...
  /scenarios/calculate:
    post:
      summary: Compute deterministic scenario preview without persisting
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScenarioCalculationRequest"
      responses:
        "200":
          description: Calculated scenario outputs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScenarioCalculationResult"
  /scenarios:
    get:
      summary: List user scenarios
      parameters:
        - name: company_id
          in: query
          schema:
            type: string
            format: uuid
        - name: document_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of saved scenarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScenarioOut"
    post:
      summary: Create and save a user scenario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScenarioCreate"
      responses:
        "201":
          description: Created scenario
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScenarioOut"
  /scenarios/{scenarioId}:
    get:
      summary: Get a specific scenario
      parameters:
        - name: scenarioId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Scenario details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScenarioOut"
    patch:
      summary: Update an existing scenario
      parameters:
        - name: scenarioId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScenarioUpdate"
      responses:
        "200":
          description: Updated scenario
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScenarioOut"
    delete:
      summary: Delete a scenario
      parameters:
        - name: scenarioId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted successfully
...
components:
  securitySchemes:
    sessionToken:
      type: http
      scheme: bearer
...
  schemas:
...
    ScenarioCalculationRequest:
      type: object
      required: [base_revenue]
      properties:
        base_revenue:
          type: string
        base_net_income:
          type: string
          nullable: true
        revenue_growth_pct:
          type: string
          default: "0.0"
        net_margin_pct:
          type: string
          nullable: true
    ScenarioCalculationResult:
      type: object
      required: [base_revenue, revenue_growth_pct, scenario_revenue, revenue_change, disclaimer]
      properties:
        base_revenue:
          type: string
        base_net_income:
          type: string
          nullable: true
        base_margin_pct:
          type: string
          nullable: true
        revenue_growth_pct:
          type: string
        net_margin_pct:
          type: string
          nullable: true
        scenario_revenue:
          type: string
        scenario_net_income:
          type: string
          nullable: true
        revenue_change:
          type: string
        net_income_change:
          type: string
          nullable: true
        disclaimer:
          type: string
    ScenarioCreate:
      type: object
      required: [name, base_period, base_revenue]
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        company_id:
          type: string
          format: uuid
          nullable: true
        document_id:
          type: string
          format: uuid
          nullable: true
        base_period:
          type: string
        base_revenue:
          type: string
        base_net_income:
          type: string
          nullable: true
        revenue_growth_pct:
          type: string
          default: "0.0"
        net_margin_pct:
          type: string
          nullable: true
        assumptions:
          type: object
    ScenarioUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        revenue_growth_pct:
          type: string
        net_margin_pct:
          type: string
          nullable: true
        assumptions:
          type: object
    ScenarioOut:
      type: object
      required: [id, user_id, name, base_period, base_revenue, revenue_growth_pct, scenario_revenue, revenue_change, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
        company_id:
          type: string
          format: uuid
          nullable: true
        document_id:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
        description:
          type: string
          nullable: true
        base_period:
          type: string
        base_revenue:
          type: string
        base_net_income:
          type: string
          nullable: true
        base_margin_pct:
          type: string
          nullable: true
        revenue_growth_pct:
          type: string
        net_margin_pct:
          type: string
          nullable: true
        scenario_revenue:
          type: string
        scenario_net_income:
          type: string
          nullable: true
        revenue_change:
          type: string
        net_income_change:
          type: string
          nullable: true
        assumptions:
          type: object
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...